Extract SSA CFG ID types into SSACFGTypes.h
